Split up base_task tests
