+++ test/integration/base_task/base_task.py
@@ -1,0 +1,3 @@
+from exasol_integration_test_docker_environment.lib.base.dependency_logger_base_task import DependencyLoggerBaseTask
+
+TestBaseTask = DependencyLoggerBaseTask
